requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
alembic==1.12.0
gunicorn==21.2.0
